wols_ca_uploader/requirements.txt
paho-mqtt>=2.1.0
pyyaml
cryptography
packaging
